Worked
I worked on a bit of the level design but decided to save the rest of it
for later when I had a little revelation.

You'll note that I added a script for moving enemies that are triggered.
And it works! So this could help make the last level and the sewer a
little more interesting. We could also use it to tweak some of the other
levels if needed.

I also fixed the sewer level. At least some, I haven't tested the
pickups and how they work in the sewer. Looking in the code I'm
concerned that they'll remove speed if they're encountered in the sewer.

This could either be a feature or be fixed rather easily. The easiest
workaround would be to just not use them in the sewer of course.